+++ src/decorators/Info.ts
@@ -1,18 +1,17 @@
 import { MetadataStorage } from "../metadata/metadata-storage";
 import { SymbolKeysNotSupportedError } from "../errors";
 
-export function Ctx(propertyName?: string): ParameterDecorator {
+export function Info(): ParameterDecorator {
   return (prototype, propertyKey, parameterIndex) => {
     if (typeof propertyKey === "symbol") {
       throw new SymbolKeysNotSupportedError();
     }
 
     MetadataStorage.registerHandlerParam({
-      kind: "context",
+      kind: "info",
       target: prototype.constructor,
       methodName: propertyKey,
       index: parameterIndex,
-      propertyName,
     });
   };
 }