+++ src/types/auth-checker.ts
@@ -2,10 +2,11 @@ import { GraphQLResolveInfo } from "graphql";
 
 export interface ActionData<ContextType = {}> {
   root: any;
-  args: any;
+  args: { [argName: string]: any };
   context: ContextType;
   info: GraphQLResolveInfo;
 }
+
 export type AuthChecker<ContextType = {}> = (
   actionData: ActionData<ContextType>,
   roles: string[],